refactor: Service를 행동에 따라 분리
CRUD (manipulatio) 를 담당하는 ReservationManipulation,
조회를 담당하는 (check)
ReservationCheck 으로 분리하였습니다.
