+++ src/main/java/com/shallwe/domain/reservation/application/ReservationManipulationServiceImpl.java
@@ -4,20 +4,16 @@ import static com.shallwe.domain.reservation.domain.ReservationStatus.WAITING;
 
 import com.shallwe.domain.common.Status;
 import com.shallwe.domain.experiencegift.domain.ExperienceGift;
-import com.shallwe.domain.experiencegift.exception.ExperienceGiftNotFoundException;
 import com.shallwe.domain.experiencegift.domain.repository.ExperienceGiftRepository;
+import com.shallwe.domain.experiencegift.exception.ExperienceGiftNotFoundException;
 import com.shallwe.domain.reservation.domain.Reservation;
 import com.shallwe.domain.reservation.domain.ReservationStatus;
 import com.shallwe.domain.reservation.domain.repository.ReservationRepository;
 import com.shallwe.domain.reservation.dto.DeleteReservationRes;
-import com.shallwe.domain.reservation.dto.ReservationIdOwnerRes;
-import com.shallwe.domain.reservation.dto.ReservationIdUserRes;
 import com.shallwe.domain.reservation.dto.ReservationRequest;
 import com.shallwe.domain.reservation.dto.ReservationResponse;
 import com.shallwe.domain.reservation.dto.ReservationUserReq;
 import com.shallwe.domain.reservation.dto.UpdateReservationReq;
-import com.shallwe.domain.reservation.dto.ValidTimeSlotRes;
-import com.shallwe.domain.reservation.exception.InvalidAvailableTimeException;
 import com.shallwe.domain.reservation.exception.InvalidReservationException;
 import com.shallwe.domain.reservation.exception.InvalidUserException;
 import com.shallwe.domain.shopowner.domain.ShopOwner;
@@ -26,68 +22,21 @@ import com.shallwe.domain.shopowner.exception.InvalidShopOwnerException;
 import com.shallwe.domain.user.domain.User;
 import com.shallwe.domain.user.domain.repository.UserRepository;
 import com.shallwe.global.config.security.token.UserPrincipal;
-import java.time.LocalDate;
-import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
-@RequiredArgsConstructor
 @Service
+@RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class ReservationServiceImpl implements ReservationService {
+public class ReservationManipulationServiceImpl implements ReservationManipulationService{
 
   private final ReservationRepository reservationRepository;
   private final ExperienceGiftRepository experienceGiftRepository;
   private final ShopOwnerRepository shopOwnerRepository;
   private final UserRepository userRepository;
-
-  public List<ValidTimeSlotRes> getValidReservationTime(UserPrincipal userPrincipal, Long giftId) {
-    ExperienceGift experienceGift = experienceGiftRepository.findById(giftId)
-        .orElseThrow(ExperienceGiftNotFoundException::new);
-
-    List<Reservation> reservations = reservationRepository.findAllByExperienceGiftAndReservationStatus(
-            experienceGift, WAITING)
-        .orElseThrow(InvalidAvailableTimeException::new);
-    System.out.println(Arrays.toString(reservations.toArray()));
-    return reservations.stream().map(reservation -> ValidTimeSlotRes.builder()
-        .reservationId(reservation.getId())
-        .date(reservation.getDate())
-        .time(reservation.getTime())
-        .build()).toList();
-  }
-
-  public List<ReservationIdUserRes> getReservationByDateUser(UserPrincipal userPrincipal, Long giftId,
-      LocalDate date) {
-    ExperienceGift experienceGift = experienceGiftRepository.findById(giftId)
-        .orElseThrow(ExperienceGiftNotFoundException::new);
-
-    List<Reservation> reservations = reservationRepository.findAllByExperienceGiftAndDate(
-            experienceGift, date)
-
-        .orElseThrow(InvalidReservationException::new);
-
-      return reservations.stream()
-              .map(ReservationIdUserRes::toDtoUser)
-              .collect(Collectors.toList());
-  }
-
-  public List<ReservationIdOwnerRes> getReservationByDateOwner(UserPrincipal userPrincipal, Long giftId,
-      LocalDate date) {
-    ExperienceGift experienceGift = experienceGiftRepository.findById(giftId)
-            .orElseThrow(ExperienceGiftNotFoundException::new);
-
-    List<Reservation> reservations = reservationRepository.findAllByExperienceGiftAndDate(
-                    experienceGift, date)
-            .orElseThrow(InvalidReservationException::new);
-
-    return reservations.stream()
-        .map(ReservationIdOwnerRes::toDtoOwner)
-        .collect(Collectors.toList());
-  }
 
   @Transactional
   public List<ReservationResponse> addOwnerReservation(ReservationRequest reservationRequest,
@@ -131,25 +80,6 @@ public class ReservationServiceImpl implements ReservationService {
     return ReservationResponse.toDtoUser(reservation);
   }
 
-  public List<ReservationResponse> findUserReservation(UserPrincipal userPrincipal) {
-    return reservationRepository.findAllBySenderId(userPrincipal.getId())
-        .stream().map(ReservationResponse::toDtoUser).collect(Collectors.toList());
-  }
-
-  public List<ReservationResponse> getAllReservation() {
-    return reservationRepository.findAll().stream().map(ReservationResponse::toDtoUser)
-        .collect(Collectors.toList());
-  }
-
-  public List<ReservationResponse> getCurrentGiftReservation(Long giftId) {
-    ExperienceGift experienceGift = experienceGiftRepository.findById(giftId)
-        .orElseThrow(ExperienceGiftNotFoundException::new);
-    List<Reservation> reservations = reservationRepository.findAllByExperienceGift(experienceGift)
-        .orElseThrow(ExperienceGiftNotFoundException::new);
-    return reservations.stream().map(ReservationResponse::toDtoUser)
-        .collect(Collectors.toList());
-  }
-
   @Transactional
   public ReservationResponse updateReservation(UpdateReservationReq updateReq,
       UserPrincipal userPrincipal) {
@@ -171,4 +101,5 @@ public class ReservationServiceImpl implements ReservationService {
     reservation.updateStatus(Status.DELETE);
     return DeleteReservationRes.toDTO();
   }
+
 }
